Implement view code, change import syntax, change html & css
Had to change the way we were importing stuff to make it work with browsers. It would have worked originally, but we would have had to build something. Using module js files are better.

Slightly altered the html & css, to improve the way we were loading the script and change the name of the display field. It didn't make sense that I had made it named "output" when it is an input field. So I changed it to "display". Not that big of a deal.